feat(security): extract security module into modules/security/ (#707)
Move 10 services, 3 routes, and pcap model into backend/src/modules/security/.
Create barrel export (index.ts) with public API and route aggregator.
Update all consumer imports: app.ts, scheduler, monitoring-service,
dashboard, monitoring, and images routes now import from module barrel.

Part of Epic #705 — Phase 2.1: first domain module extraction.
