+++ backend/src/modules/security/routes/pcap.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import { FastifyInstance } from 'fastify';
-import { StartCaptureRequestSchema, CaptureListQuerySchema } from '../core/models/pcap.js';
-import { ActionIdParamsSchema } from '../core/models/api-schemas.js';
+import { StartCaptureRequestSchema, CaptureListQuerySchema } from '../models/pcap.js';
+import { ActionIdParamsSchema } from '../../../core/models/api-schemas.js';
 import {
   startCapture,
   stopCapture,
@@ -11,9 +11,9 @@ import {
   getCaptureFilePath,
 } from '../services/pcap-service.js';
 import { analyzeCapture } from '../services/pcap-analysis-service.js';
-import { writeAuditLog } from '../core/services/audit-logger.js';
-import { assertCapability } from '../services/edge-capability-guard.js';
-import { createChildLogger } from '../core/utils/logger.js';
+import { writeAuditLog } from '../../../core/services/audit-logger.js';
+import { assertCapability } from '../../../services/edge-capability-guard.js';
+import { createChildLogger } from '../../../core/utils/logger.js';
 
 const log = createChildLogger('pcap-route');
