+++ backend/src/modules/security/services/harbor-vulnerability-store.ts
@@ -1,5 +1,5 @@
-import { getDbForDomain } from '../core/db/app-db-router.js';
-import { createChildLogger } from '../core/utils/logger.js';
+import { getDbForDomain } from '../../../core/db/app-db-router.js';
+import { createChildLogger } from '../../../core/utils/logger.js';
 
 const log = createChildLogger('harbor-vulnerability-store');
