+++ backend/src/modules/security/services/pcap-store.ts
@@ -1,6 +1,6 @@
-import { getDbForDomain } from '../core/db/app-db-router.js';
-import { createChildLogger } from '../core/utils/logger.js';
-import type { Capture, CaptureStatus } from '../core/models/pcap.js';
+import { getDbForDomain } from '../../../core/db/app-db-router.js';
+import { createChildLogger } from '../../../core/utils/logger.js';
+import type { Capture, CaptureStatus } from '../models/pcap.js';
 
 const log = createChildLogger('pcap-store');
