+++ backend/src/modules/security/services/security-scanner.ts
@@ -1,5 +1,5 @@
-import { createChildLogger } from '../core/utils/logger.js';
-import type { Container } from '../core/models/portainer.js';
+import { createChildLogger } from '../../../core/utils/logger.js';
+import type { Container } from '../../../core/models/portainer.js';
 
 const log = createChildLogger('security-scanner');
