+++ backend/src/routes/dashboard.ts
@@ -5,7 +5,7 @@ import { cachedFetchSWR, getCacheKey, TTL } from '../core/portainer/portainer-ca
 import { normalizeEndpoint, normalizeContainer } from '../core/portainer/portainer-normalizers.js';
 import { getKpiHistory } from '../services/kpi-store.js';
 import { createChildLogger } from '../core/utils/logger.js';
-import { buildSecurityAuditSummary, getSecurityAudit } from '../services/security-audit.js';
+import { buildSecurityAuditSummary, getSecurityAudit } from '../modules/security/index.js';
 import { getLatestMetricsBatch } from '../services/metrics-store.js';
 
 const log = createChildLogger('route:dashboard');
