+++ backend/src/routes/images.ts
@@ -2,7 +2,7 @@ import { FastifyInstance } from 'fastify';
 import * as portainer from '../core/portainer/portainer-client.js';
 import { cachedFetch, getCacheKey, TTL } from '../core/portainer/portainer-cache.js';
 import { EndpointIdQuerySchema } from '../core/models/api-schemas.js';
-import { getStalenessRecords, getStalenessSummary, runStalenessChecks } from '../services/image-staleness.js';
+import { getStalenessRecords, getStalenessSummary, runStalenessChecks } from '../modules/security/index.js';
 import { createChildLogger } from '../core/utils/logger.js';
 
 const log = createChildLogger('route:images');
